Fix PRIME Sync/Offload Configuration Issue
As per the NixOS wiki, PRIME sync and Offload **cannot** be enaabled at the same time. I have assumed the user will want to have less screen tearing and disabled offloading. You could do it the other way (or even use `reverseSync = true;` if assuming a newer driver and GPU), whatever works but as of recent nixpkgs-unstable builds, this would throw an error and not build as it is.